+++ Powder/GOL.h
@@ -24,7 +24,7 @@ public:
 	// Processes the rule string and fills the rules array
 	void process_rules();
 	bool update(double dt) override;
-	void render(float, float, sf::Vertex* quad) override;
+	void render(double, double, sf::Vertex* quad) override;
 	Element* clone() const override;
 	GOL(int, std::string, std::string);
 	GOL(const GOL& rhs);
